test(mcp-doctor): split artifact detection coverage [砚砚/GPT-5.4🐾]
